refactor: code optimization
just discovered that it is possible to name a for loop. shocking.
change isTouched() method's parameter to float for consistency

+++ app/src/main/java/imo/text/Line.java
@@ -16,7 +16,7 @@ public class Line {
         this.text = text;
     }
     
-    boolean isTouched(int touchY){
+    boolean isTouched(float touchY){
         return touchY <= bottom && touchY >= top;
     }
 }
